[DoctrineBridge] Remove DoctrineTestHelper and TestRepositoryFactory
